very little refactoring of parser.h
