+++ src/parse_instRW.h
@@ -1,0 +1,210 @@
+#ifndef VIRTUAL_PARSE_INSTRW_H
+#define VIRTUAL_PARSE_INSTRW_H
+
+#include "core.h"
+
+static inline int __cmp_str_view(const char* sv_cstr, int size, const char* cstr){
+
+    int i = 0;
+
+    for(; i < size && sv_cstr[i] == cstr[i]; i+=1);
+
+    return i == size && !cstr[i];
+}
+
+uint8_t _get_inst_op(const char* inst_cstr, int size){
+    #define __VIRTUAL_RETURN_INST_OP(INST_NAME, OP_PROFILE, ...)\
+        if(__cmp_str_view(inst_cstr, size, #INST_NAME)) return INST_ ## INST_NAME
+
+    __VIRTUAL_RETURN_INST_OP(NOP,       OP_PROFILE_NONE, INST_NOP  );
+    __VIRTUAL_RETURN_INST_OP(HALT,      OP_PROFILE_E  , INST_HALT  );
+    __VIRTUAL_RETURN_INST_OP(MOV8,      OP_PROFILE_RR , INST_MOV8  );
+    __VIRTUAL_RETURN_INST_OP(MOV16,     OP_PROFILE_RR , INST_MOV16 );
+    __VIRTUAL_RETURN_INST_OP(MOV32,     OP_PROFILE_RR , INST_MOV32 );
+    __VIRTUAL_RETURN_INST_OP(MOV,       OP_PROFILE_RR , INST_MOV   );
+    __VIRTUAL_RETURN_INST_OP(MOVC,      OP_PROFILE_RRR, INST_MOVC  );
+    __VIRTUAL_RETURN_INST_OP(MOVV,      OP_PROFILE_RL , INST_MOVV  );
+    __VIRTUAL_RETURN_INST_OP(MOVN,      OP_PROFILE_RL , INST_MOVN  );
+    __VIRTUAL_RETURN_INST_OP(MOVV16,    OP_PROFILE_RL , INST_MOVV16);
+    __VIRTUAL_RETURN_INST_OP(PUSH,      OP_PROFILE_E  , INST_PUSH  );
+    __VIRTUAL_RETURN_INST_OP(POP,       OP_PROFILE_R  , INST_POP   );
+    __VIRTUAL_RETURN_INST_OP(STACK_GET, OP_PROFILE_RL , INST_STACK_GET);
+    __VIRTUAL_RETURN_INST_OP(STACK_PUT, OP_PROFILE_RL , INST_STACK_PUT);
+    __VIRTUAL_RETURN_INST_OP(GSP,       OP_PROFILE_RRR, INST_GSP   );
+    __VIRTUAL_RETURN_INST_OP(STATIC,    OP_PROFILE_E  , INST_STATIC);
+    __VIRTUAL_RETURN_INST_OP(READ8,     OP_PROFILE_RRR, INST_READ8 );
+    __VIRTUAL_RETURN_INST_OP(READ16,    OP_PROFILE_RRR, INST_READ16);
+    __VIRTUAL_RETURN_INST_OP(READ32,    OP_PROFILE_RRR, INST_READ32);
+    __VIRTUAL_RETURN_INST_OP(READ,      OP_PROFILE_RRR, INST_READ  );
+    __VIRTUAL_RETURN_INST_OP(WRITE8,    OP_PROFILE_RRR, INST_WRITE8 );
+    __VIRTUAL_RETURN_INST_OP(WRITE16,   OP_PROFILE_RRR, INST_WRITE16);
+    __VIRTUAL_RETURN_INST_OP(WRITE32,   OP_PROFILE_RRR, INST_WRITE32);
+    __VIRTUAL_RETURN_INST_OP(WRITE,     OP_PROFILE_RRR, INST_WRITE  );
+    __VIRTUAL_RETURN_INST_OP(MWRITES,   OP_PROFILE_RRR, INST_MWRITES);
+    __VIRTUAL_RETURN_INST_OP(MMOVS,     OP_PROFILE_RRR, INST_MMOVS );
+    __VIRTUAL_RETURN_INST_OP(MEMCMP,    OP_PROFILE_RRR, INST_MEMCMP);
+    __VIRTUAL_RETURN_INST_OP(NOT,       OP_PROFILE_RR , INST_NOT   );
+    __VIRTUAL_RETURN_INST_OP(NEG,       OP_PROFILE_RRR, INST_NEG   );
+    __VIRTUAL_RETURN_INST_OP(AND,       OP_PROFILE_RRR, INST_AND   );
+    __VIRTUAL_RETURN_INST_OP(NAND,      OP_PROFILE_RRR, INST_NAND  );
+    __VIRTUAL_RETURN_INST_OP(OR,        OP_PROFILE_RRR, INST_OR    );
+    __VIRTUAL_RETURN_INST_OP(XOR,       OP_PROFILE_RRR, INST_XOR   );
+    __VIRTUAL_RETURN_INST_OP(BSHIFT,    OP_PROFILE_RRR, INST_BSHIFT);
+    __VIRTUAL_RETURN_INST_OP(JMP,       OP_PROFILE_E  , INST_JMP   );
+    __VIRTUAL_RETURN_INST_OP(JMPF,      OP_PROFILE_RL , INST_JMPF  );
+    __VIRTUAL_RETURN_INST_OP(JMPFN,     OP_PROFILE_RL , INST_JMPFN );
+    __VIRTUAL_RETURN_INST_OP(CALL,      OP_PROFILE_E  , INST_CALL  );
+    __VIRTUAL_RETURN_INST_OP(RET,       OP_PROFILE_NONE, INST_RET  );
+    __VIRTUAL_RETURN_INST_OP(ADD8,      OP_PROFILE_RRR, INST_ADD8  );
+    __VIRTUAL_RETURN_INST_OP(SUB8,      OP_PROFILE_RRR, INST_SUB8  );
+    __VIRTUAL_RETURN_INST_OP(MUL8,      OP_PROFILE_RRR, INST_MUL8  );
+    __VIRTUAL_RETURN_INST_OP(ADD16,     OP_PROFILE_RRR, INST_ADD16 );
+    __VIRTUAL_RETURN_INST_OP(SUB16,     OP_PROFILE_RRR, INST_SUB16 );
+    __VIRTUAL_RETURN_INST_OP(MUL16,     OP_PROFILE_RRR, INST_MUL16 );
+    __VIRTUAL_RETURN_INST_OP(ADD32,     OP_PROFILE_RRR, INST_ADD32 );
+    __VIRTUAL_RETURN_INST_OP(SUB32,     OP_PROFILE_RRR, INST_SUB32 );
+    __VIRTUAL_RETURN_INST_OP(MUL32,     OP_PROFILE_RRR, INST_MUL32 );
+    __VIRTUAL_RETURN_INST_OP(ADD,       OP_PROFILE_RRR, INST_ADD   );
+    __VIRTUAL_RETURN_INST_OP(SUB,       OP_PROFILE_RRR, INST_SUB   );
+    __VIRTUAL_RETURN_INST_OP(MUL,       OP_PROFILE_RRR, INST_MUL   );
+    __VIRTUAL_RETURN_INST_OP(DIVI,      OP_PROFILE_RRR, INST_DIVI  );
+    __VIRTUAL_RETURN_INST_OP(DIVU,      OP_PROFILE_RRR, INST_DIVU  );
+    __VIRTUAL_RETURN_INST_OP(ADDF,      OP_PROFILE_RRR, INST_ADDF  );
+    __VIRTUAL_RETURN_INST_OP(SUBF,      OP_PROFILE_RRR, INST_SUBF  );
+    __VIRTUAL_RETURN_INST_OP(MULF,      OP_PROFILE_RRR, INST_MULF  );
+    __VIRTUAL_RETURN_INST_OP(DIVF,      OP_PROFILE_RRR, INST_DIVF  );
+    __VIRTUAL_RETURN_INST_OP(INC,       OP_PROFILE_RL , INST_INC   );
+    __VIRTUAL_RETURN_INST_OP(DEC,       OP_PROFILE_RL , INST_DEC   );
+    __VIRTUAL_RETURN_INST_OP(INCF,      OP_PROFILE_RL , INST_INCF  );
+    __VIRTUAL_RETURN_INST_OP(DECF,      OP_PROFILE_RL , INST_DECF  );
+    __VIRTUAL_RETURN_INST_OP(ABS,       OP_PROFILE_RRR, INST_ABS   );
+    __VIRTUAL_RETURN_INST_OP(ABSF,      OP_PROFILE_RRR, INST_ABSF  );
+    __VIRTUAL_RETURN_INST_OP(NEQ,       OP_PROFILE_RRR, INST_NEQ   );
+    __VIRTUAL_RETURN_INST_OP(EQ,        OP_PROFILE_RRR, INST_EQ    );
+    __VIRTUAL_RETURN_INST_OP(EQF,       OP_PROFILE_RRR, INST_EQF   );
+    __VIRTUAL_RETURN_INST_OP(BIGI,      OP_PROFILE_RRR, INST_BIGI  );
+    __VIRTUAL_RETURN_INST_OP(BIGU,      OP_PROFILE_RRR, INST_BIGU  );
+    __VIRTUAL_RETURN_INST_OP(BIGF,      OP_PROFILE_RRR, INST_BIGF  );
+    __VIRTUAL_RETURN_INST_OP(SMLI,      OP_PROFILE_RRR, INST_SMLI  );
+    __VIRTUAL_RETURN_INST_OP(SMLU,      OP_PROFILE_RRR, INST_SMLU  );
+    __VIRTUAL_RETURN_INST_OP(SMLF,      OP_PROFILE_RRR, INST_SMLF  );
+    __VIRTUAL_RETURN_INST_OP(CASTIU,    OP_PROFILE_RR , INST_CASTIU);
+    __VIRTUAL_RETURN_INST_OP(CASTIF,    OP_PROFILE_RR , INST_CASTIF);
+    __VIRTUAL_RETURN_INST_OP(CASTUI,    OP_PROFILE_RR , INST_CASTUI);
+    __VIRTUAL_RETURN_INST_OP(CASTUF,    OP_PROFILE_RR , INST_CASTUF);
+    __VIRTUAL_RETURN_INST_OP(CASTFI,    OP_PROFILE_RR , INST_CASTFI);
+    __VIRTUAL_RETURN_INST_OP(CASTFU,    OP_PROFILE_RR , INST_CASTFU);
+    __VIRTUAL_RETURN_INST_OP(CF3264,    OP_PROFILE_RR , INST_CF3264);
+    __VIRTUAL_RETURN_INST_OP(CF6432,    OP_PROFILE_RR , INST_CF6432);
+    __VIRTUAL_RETURN_INST_OP(FLOAT,     OP_PROFILE_RRR, INST_FLOAT );
+    __VIRTUAL_RETURN_INST_OP(DUMPCHAR,  OP_PROFILE_RRR, INST_DUMPCHAR);
+    __VIRTUAL_RETURN_INST_OP(GETCHAR,   OP_PROFILE_RR , INST_GETCHAR);
+    __VIRTUAL_RETURN_INST_OP(EXEC,      OP_PROFILE_R  , INST_EXEC  );
+    __VIRTUAL_RETURN_INST_OP(SYS,       OP_PROFILE_E  , INST_SYS   );
+    __VIRTUAL_RETURN_INST_OP(DISREG,    OP_PROFILE_RRR, INST_DISREG);
+
+    return INST_ERROR;
+    #undef __VIRTUAL_RETURN_INST_OP
+}
+
+OpProfile get_inst_profile(int inst){
+    #define __VIRTUAL_RETURN_INST_PROFILE(INST_NAME, OP_PROFILE, ...)\
+        case INST_ ## INST_NAME: return OP_PROFILE
+
+    switch (inst)
+    {
+    __VIRTUAL_RETURN_INST_PROFILE(NOP,       OP_PROFILE_NONE, INST_NOP  );
+    __VIRTUAL_RETURN_INST_PROFILE(HALT,      OP_PROFILE_E  , INST_HALT  );
+    __VIRTUAL_RETURN_INST_PROFILE(MOV8,      OP_PROFILE_RR , INST_MOV8  );
+    __VIRTUAL_RETURN_INST_PROFILE(MOV16,     OP_PROFILE_RR , INST_MOV16 );
+    __VIRTUAL_RETURN_INST_PROFILE(MOV32,     OP_PROFILE_RR , INST_MOV32 );
+    __VIRTUAL_RETURN_INST_PROFILE(MOV,       OP_PROFILE_RR , INST_MOV   );
+    __VIRTUAL_RETURN_INST_PROFILE(MOVC,      OP_PROFILE_RRR, INST_MOVC  );
+    __VIRTUAL_RETURN_INST_PROFILE(MOVV,      OP_PROFILE_RL , INST_MOVV  );
+    __VIRTUAL_RETURN_INST_PROFILE(MOVN,      OP_PROFILE_RL , INST_MOVN  );
+    __VIRTUAL_RETURN_INST_PROFILE(MOVV16,    OP_PROFILE_RL , INST_MOVV16);
+    __VIRTUAL_RETURN_INST_PROFILE(PUSH,      OP_PROFILE_E  , INST_PUSH  );
+    __VIRTUAL_RETURN_INST_PROFILE(POP,       OP_PROFILE_R  , INST_POP   );
+    __VIRTUAL_RETURN_INST_PROFILE(STACK_GET, OP_PROFILE_RL , INST_STACK_GET);
+    __VIRTUAL_RETURN_INST_PROFILE(STACK_PUT, OP_PROFILE_RL , INST_STACK_PUT);
+    __VIRTUAL_RETURN_INST_PROFILE(GSP,       OP_PROFILE_RRR, INST_GSP   );
+    __VIRTUAL_RETURN_INST_PROFILE(STATIC,    OP_PROFILE_E  , INST_STATIC);
+    __VIRTUAL_RETURN_INST_PROFILE(READ8,     OP_PROFILE_RRR, INST_READ8 );
+    __VIRTUAL_RETURN_INST_PROFILE(READ16,    OP_PROFILE_RRR, INST_READ16);
+    __VIRTUAL_RETURN_INST_PROFILE(READ32,    OP_PROFILE_RRR, INST_READ32);
+    __VIRTUAL_RETURN_INST_PROFILE(READ,      OP_PROFILE_RRR, INST_READ  );
+    __VIRTUAL_RETURN_INST_PROFILE(WRITE8,    OP_PROFILE_RRR, INST_WRITE8 );
+    __VIRTUAL_RETURN_INST_PROFILE(WRITE16,   OP_PROFILE_RRR, INST_WRITE16);
+    __VIRTUAL_RETURN_INST_PROFILE(WRITE32,   OP_PROFILE_RRR, INST_WRITE32);
+    __VIRTUAL_RETURN_INST_PROFILE(WRITE,     OP_PROFILE_RRR, INST_WRITE  );
+    __VIRTUAL_RETURN_INST_PROFILE(MWRITES,   OP_PROFILE_RRR, INST_MWRITES);
+    __VIRTUAL_RETURN_INST_PROFILE(MMOVS,     OP_PROFILE_RRR, INST_MMOVS );
+    __VIRTUAL_RETURN_INST_PROFILE(MEMCMP,    OP_PROFILE_RRR, INST_MEMCMP);
+    __VIRTUAL_RETURN_INST_PROFILE(NOT,       OP_PROFILE_RR , INST_NOT   );
+    __VIRTUAL_RETURN_INST_PROFILE(NEG,       OP_PROFILE_RRR, INST_NEG   );
+    __VIRTUAL_RETURN_INST_PROFILE(AND,       OP_PROFILE_RRR, INST_AND   );
+    __VIRTUAL_RETURN_INST_PROFILE(NAND,      OP_PROFILE_RRR, INST_NAND  );
+    __VIRTUAL_RETURN_INST_PROFILE(OR,        OP_PROFILE_RRR, INST_OR    );
+    __VIRTUAL_RETURN_INST_PROFILE(XOR,       OP_PROFILE_RRR, INST_XOR   );
+    __VIRTUAL_RETURN_INST_PROFILE(BSHIFT,    OP_PROFILE_RRR, INST_BSHIFT);
+    __VIRTUAL_RETURN_INST_PROFILE(JMP,       OP_PROFILE_E  , INST_JMP   );
+    __VIRTUAL_RETURN_INST_PROFILE(JMPF,      OP_PROFILE_RL , INST_JMPF  );
+    __VIRTUAL_RETURN_INST_PROFILE(JMPFN,     OP_PROFILE_RL , INST_JMPFN );
+    __VIRTUAL_RETURN_INST_PROFILE(CALL,      OP_PROFILE_E  , INST_CALL  );
+    __VIRTUAL_RETURN_INST_PROFILE(RET,       OP_PROFILE_NONE, INST_RET  );
+    __VIRTUAL_RETURN_INST_PROFILE(ADD8,      OP_PROFILE_RRR, INST_ADD8  );
+    __VIRTUAL_RETURN_INST_PROFILE(SUB8,      OP_PROFILE_RRR, INST_SUB8  );
+    __VIRTUAL_RETURN_INST_PROFILE(MUL8,      OP_PROFILE_RRR, INST_MUL8  );
+    __VIRTUAL_RETURN_INST_PROFILE(ADD16,     OP_PROFILE_RRR, INST_ADD16 );
+    __VIRTUAL_RETURN_INST_PROFILE(SUB16,     OP_PROFILE_RRR, INST_SUB16 );
+    __VIRTUAL_RETURN_INST_PROFILE(MUL16,     OP_PROFILE_RRR, INST_MUL16 );
+    __VIRTUAL_RETURN_INST_PROFILE(ADD32,     OP_PROFILE_RRR, INST_ADD32 );
+    __VIRTUAL_RETURN_INST_PROFILE(SUB32,     OP_PROFILE_RRR, INST_SUB32 );
+    __VIRTUAL_RETURN_INST_PROFILE(MUL32,     OP_PROFILE_RRR, INST_MUL32 );
+    __VIRTUAL_RETURN_INST_PROFILE(ADD,       OP_PROFILE_RRR, INST_ADD   );
+    __VIRTUAL_RETURN_INST_PROFILE(SUB,       OP_PROFILE_RRR, INST_SUB   );
+    __VIRTUAL_RETURN_INST_PROFILE(MUL,       OP_PROFILE_RRR, INST_MUL   );
+    __VIRTUAL_RETURN_INST_PROFILE(DIVI,      OP_PROFILE_RRR, INST_DIVI  );
+    __VIRTUAL_RETURN_INST_PROFILE(DIVU,      OP_PROFILE_RRR, INST_DIVU  );
+    __VIRTUAL_RETURN_INST_PROFILE(ADDF,      OP_PROFILE_RRR, INST_ADDF  );
+    __VIRTUAL_RETURN_INST_PROFILE(SUBF,      OP_PROFILE_RRR, INST_SUBF  );
+    __VIRTUAL_RETURN_INST_PROFILE(MULF,      OP_PROFILE_RRR, INST_MULF  );
+    __VIRTUAL_RETURN_INST_PROFILE(DIVF,      OP_PROFILE_RRR, INST_DIVF  );
+    __VIRTUAL_RETURN_INST_PROFILE(INC,       OP_PROFILE_RL , INST_INC   );
+    __VIRTUAL_RETURN_INST_PROFILE(DEC,       OP_PROFILE_RL , INST_DEC   );
+    __VIRTUAL_RETURN_INST_PROFILE(INCF,      OP_PROFILE_RL , INST_INCF  );
+    __VIRTUAL_RETURN_INST_PROFILE(DECF,      OP_PROFILE_RL , INST_DECF  );
+    __VIRTUAL_RETURN_INST_PROFILE(ABS,       OP_PROFILE_RRR, INST_ABS   );
+    __VIRTUAL_RETURN_INST_PROFILE(ABSF,      OP_PROFILE_RRR, INST_ABSF  );
+    __VIRTUAL_RETURN_INST_PROFILE(NEQ,       OP_PROFILE_RRR, INST_NEQ   );
+    __VIRTUAL_RETURN_INST_PROFILE(EQ,        OP_PROFILE_RRR, INST_EQ    );
+    __VIRTUAL_RETURN_INST_PROFILE(EQF,       OP_PROFILE_RRR, INST_EQF   );
+    __VIRTUAL_RETURN_INST_PROFILE(BIGI,      OP_PROFILE_RRR, INST_BIGI  );
+    __VIRTUAL_RETURN_INST_PROFILE(BIGU,      OP_PROFILE_RRR, INST_BIGU  );
+    __VIRTUAL_RETURN_INST_PROFILE(BIGF,      OP_PROFILE_RRR, INST_BIGF  );
+    __VIRTUAL_RETURN_INST_PROFILE(SMLI,      OP_PROFILE_RRR, INST_SMLI  );
+    __VIRTUAL_RETURN_INST_PROFILE(SMLU,      OP_PROFILE_RRR, INST_SMLU  );
+    __VIRTUAL_RETURN_INST_PROFILE(SMLF,      OP_PROFILE_RRR, INST_SMLF  );
+    __VIRTUAL_RETURN_INST_PROFILE(CASTIU,    OP_PROFILE_RR , INST_CASTIU);
+    __VIRTUAL_RETURN_INST_PROFILE(CASTIF,    OP_PROFILE_RR , INST_CASTIF);
+    __VIRTUAL_RETURN_INST_PROFILE(CASTUI,    OP_PROFILE_RR , INST_CASTUI);
+    __VIRTUAL_RETURN_INST_PROFILE(CASTUF,    OP_PROFILE_RR , INST_CASTUF);
+    __VIRTUAL_RETURN_INST_PROFILE(CASTFI,    OP_PROFILE_RR , INST_CASTFI);
+    __VIRTUAL_RETURN_INST_PROFILE(CASTFU,    OP_PROFILE_RR , INST_CASTFU);
+    __VIRTUAL_RETURN_INST_PROFILE(CF3264,    OP_PROFILE_RR , INST_CF3264);
+    __VIRTUAL_RETURN_INST_PROFILE(CF6432,    OP_PROFILE_RR , INST_CF6432);
+    __VIRTUAL_RETURN_INST_PROFILE(FLOAT,     OP_PROFILE_RRR, INST_FLOAT );
+    __VIRTUAL_RETURN_INST_PROFILE(DUMPCHAR,  OP_PROFILE_RRR, INST_DUMPCHAR);
+    __VIRTUAL_RETURN_INST_PROFILE(GETCHAR,   OP_PROFILE_RR , INST_GETCHAR);
+    __VIRTUAL_RETURN_INST_PROFILE(EXEC,      OP_PROFILE_R  , INST_EXEC  );
+    __VIRTUAL_RETURN_INST_PROFILE(SYS,       OP_PROFILE_E  , INST_SYS   );
+    __VIRTUAL_RETURN_INST_PROFILE(DISREG,    OP_PROFILE_RRR, INST_DISREG);
+    default:
+        return OP_PROFILE_NONE;
+    }
+
+    return OP_PROFILE_NONE;
+    #undef __VIRTUAL_RETURN_INST_PROFILE
+}
+
+#endif // =====================  END OF FILE VIRTUAL_PARSE_INSTRW_H ===========================
